+++ libs/constants/release_notes.md
@@ -4,6 +4,11 @@
 
 * [new] Add HOST_SYNC_TOOL_CONFIG_TOPIC constant.
 
+## v3.0.0
+2024-09-19
+
+* [changed] Update ftrack utils dependency to 3.0.0.
+
 
 ## v2.0.0
 2024-02-12