+++ libs/qt/release_notes.md
@@ -7,6 +7,12 @@
 * [changed] Improved AccordionWidget and AcordionHeaderWidget to use EditableLabelWidget and possibility to add a remove button. Emits bin_clicked, title_changed, title_edited signals.
 * [changed] StatusMaterialIconWidget clickable.
 * [new] Editable label widget.
+
+
+## v3.0.0
+2024-09-19
+
+* [changed] Update framework dependencies to 3.0.0.
 * [changed] FileDialog; support selecting multiple files.
 * [changed] FileBrowser; support collection of multiple files.
 * [changed] Python clique 1.6.1 library as dependency.